+++ CHANGELOG.md
@@ -10,11 +10,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - test level up on local to ensure it triggers therapy cost adjustment
 - test login on local to ensure it triggers therapy cost adjustment
 - update localization for amnesia buffs (instruct player to seek treatment from traders, maybe even show treatment cost)
-- update amn help command to return info around `buffAmnesiaFragileMemory` instead of hardened memory
 
 ## [UNRELEASED]
 
 - add dialog purchasing system for treatments
+- add success/unnecessary/unaffordable responses
 - remove deprecated buff
 - update therapy prices on login and level-up